requirements.txt
pandas==1.2.4
boto3==1.17.91
uuid
twint
spacy==3.0.6
bs4==0.0.1
pymongo==3.11.4
urllib3==1.26.4
requests==2.25.1
wikipedia==1.4.0
urlexpander
moviepy==1.0.3
pydub==0.25.1
azure.cognitiveservices.speech
azure==4.0.0
rev_ai
Pillow==8.2.0
numpy==1.19.5
nltk==3.6.2
ibm_watson
opencv-python==4.5.1.48